DiceRoller-WoD working.
The DiceRoller is in a working state. He has options for simulating an
instant action and extended actions. For the extended actions, the program has the possibility to check for a target value or checking the number of successes the Dice turn up, after a specified number of rolls.

There is even the possibility to turn on a houserule-mode. In this mode,
the DiceRoller-WoD calculates exceptional successes differently than in
the official rules from the World of Darkness.