Fixing the display of the result quality.
When rolling a specified number of rolls the result quality was wrongly
displayed. Whenever the roll accumulated more than 5 successes, this was
seen as an exceptional succes. Now, an exceptional success is
shown only then, if at least one roll in the whole extended roll has achieved five
or more successes.