Switching from PyQt4 to PySide.
This project switches from PyQt to PySide due to its no as rigid license
modell. A few changes in the codebase were necessary in the process, but
now the DiceRoller is working again.

Not connected to the library change, the display of a "rolling" die will
be changed. As of now, no die is shown at all, but this will change
soon.